+++ src/__tests__/aggregatorDoneThenWait.js
@@ -1,0 +1,6 @@
+export default function start(step) {
+  return Promise
+    .resolve()
+    .then(step("done soon"))
+    .then(() => "finished");
+}